README.md
This is my personal GitHub profile README, where I introduce myself and showcase my journey as a developer. As a beginner in Python, I'm eager to learn and grow into a Full Stack Developer and AI enthusiast. In this README, you'll find:

A little bit about me and my tech interests 🎯

The current technologies I'm working with and planning to learn 💻
Links to connect with me professionally 🌐

Follow me to stay updated on my latest projects as I work on exciting new ideas and explore the world of software development!